SortHistos.C and SortTrees.C both compiling again, finally! SortTrees.C not changed here, lots of work on SortHistos.C:
- New functions created to configure fit settings, ConfigureEnergyFit() and ConfigureWaveEnFit();
- Lots of code moved from Calibration/Fit to these new functions.
- Full set of fitting/calibration functions now set up.
   - CalibrateFiles() which calls
      - FitHistoFile() which calls first
         - ConfigureEnergyFit() and then..
         - FitGammaSpectrum() which IDs peaks and then calls
            - FitSinglePeak() which then passes back all the way to CalibrateFiles() which loops channels and calls
      - CalibrateChannel()

The CalibrateGammaSpectrum() function in CalibTools.C is now deprecated but left in as SortTrees.C still calls the old method.  This will be fixed next.

So compiling is now fine, basic code is all there, but it's not running properly just quits without doing calibration.  Surely a relatively simple bug, going to look at that next.